+++ src/GRA.Domain.Model/Filters/UserFilter.cs
@@ -1,17 +1,16 @@
-﻿
-namespace GRA.Domain.Model.Filters
+﻿namespace GRA.Domain.Model.Filters
 {
     public class UserFilter : BaseFilter
     {
-        public SortUsersBy SortBy { get; set; }
-        public bool OrderDescending { get; set; }
-        public bool CanAddToHousehold { get; set; }
-
-        public bool? IsSubscribed { get; set; }
-
         public UserFilter(int? page = null) : base(page)
         {
             SortBy = SortUsersBy.LastName;
         }
+
+        public bool CanAddToHousehold { get; set; }
+        public bool? HasMultiplePrimaryVendorCodes { get; set; }
+        public bool? IsSubscribed { get; set; }
+        public bool OrderDescending { get; set; }
+        public SortUsersBy SortBy { get; set; }
     }
 }